+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Bug fixes:
 * Fix `String#[Regexp, Integer]` when the capture group exists but is not matched (@eregon).
 * Fix `File.open` mode string parsing when binary option is the third character (@bjfish).
 * Fix `rb_scan_args_kw` macro to avoid shadowing variables (#2649, @aardvark179).
+* Fix `Float#round` to avoid losing precision during the rounding process (@aardvark179).
 
 Compatibility: